Added framerate counter and decided to not lock the fps because I am very very lazy and that work is BORRRRINGG (-right now. I'll do it later)

+++ VertexCortex Purified/Window.cpp
@@ -203,7 +203,7 @@ void Clock::tick()
 
     QueryPerformanceCounter(&currentTime);
 
-    dt = (currentTime.QuadPart - lastTime.QuadPart) / counterFrequency.QuadPart;
+    dt = (float)(currentTime.QuadPart - lastTime.QuadPart) / (float)counterFrequency.QuadPart;
 
     lastTime = currentTime;
 }
